+++ src/backends/maildir/maildir_mbox.rs
@@ -19,13 +19,16 @@ use crate::{
 
 /// Represents a list of Maildir mailboxes.
 #[derive(Debug, Default, serde::Serialize)]
-pub struct MaildirMboxes(pub Vec<MaildirMbox>);
+pub struct MaildirMboxes {
+    #[serde(rename = "response")]
+    pub mboxes: Vec<MaildirMbox>,
+}
 
 impl Deref for MaildirMboxes {
     type Target = Vec<MaildirMbox>;
 
     fn deref(&self) -> &Self::Target {
-        &self.0
+        &self.mboxes
     }
 }
 
@@ -113,7 +116,7 @@ impl TryFrom<RawMaildirMboxes> for MaildirMboxes {
         for entry in mail_entries {
             mboxes.push(entry?.try_into()?);
         }
-        Ok(MaildirMboxes(mboxes))
+        Ok(MaildirMboxes { mboxes })
     }
 }