+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `In-Reply-To` not set properly when replying to a message [#323]
 - `Cc` missing or invalid when replying to a message [#324]
 - Notmuch backend hangs [#329]
+- Maildir e2e tests [#335]
+- JSON API for listings [#331]
 
 ## [0.5.8] - 2022-03-04
 
@@ -499,3 +501,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#323]: https://github.com/soywod/himalaya/issues/323
 [#324]: https://github.com/soywod/himalaya/issues/324
 [#329]: https://github.com/soywod/himalaya/issues/329
+[#331]: https://github.com/soywod/himalaya/issues/331
+[#335]: https://github.com/soywod/himalaya/issues/335